requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
aniso8601==9.0.1
annotated-types==0.7.0
APScheduler==3.11.0
attrs==24.3.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.1.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cryptography==44.0.0
cymem==2.0.10
dnspython==2.7.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
Mako==1.3.8
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.11
nltk==3.9.1
numpy==2.2.1
packaging==24.2
PGPy==0.6.0
preshed==3.0.9
propcache==0.2.1
protonmail-api-client==2.0.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyMySQL==1.1.1
pyodbc==5.2.0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.5.0
thinc==8.3.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.0
WTForms==3.2.1
yarl==1.18.3
